LPS-78225 Move journal.transformer into internal
