+++ modules/apps/web-experience/journal/journal-service/src/main/resources/META-INF/module-log4j.xml
@@ -6,15 +6,15 @@
 		<priority value="ERROR" />
 	</category>
 
+	<category name="com.liferay.journal.internal.transformer.RegexTransformerListener">
+		<priority value="INFO" />
+	</category>
+
 	<category name="com.liferay.journal.lar">
 		<priority value="WARN" />
 	</category>
 
 	<category name="com.liferay.journal.service.impl.JournalContentSearchLocalServiceImpl">
-		<priority value="INFO" />
-	</category>
-
-	<category name="com.liferay.journal.transformer.RegexTransformerListener">
 		<priority value="INFO" />
 	</category>
