+++ modules/apps/web-experience/journal/journal-test/src/testIntegration/java/com/liferay/journal/transformer/test/JournalTransformerTest.java
@@ -356,7 +356,8 @@ public class JournalTransformerTest {
 	protected void initRegexTransformerListener() {
 		TransformerListener transformerListener =
 			_journalTransformerListenerRegistry.getTransformerListener(
-				"com.liferay.journal.transformer.RegexTransformerListener");
+				"com.liferay.journal.internal.transformer." +
+					"RegexTransformerListener");
 
 		CacheRegistryUtil.setActive(true);
